requirements.txt
python-telegram-bot==21.6
xmltodict==0.14.2
python-dotenv==1.0.0
requests==2.32.3